Struct linear_regression::save::Store

pub struct Store {
    // some fields omitted
}

The Store stucture is an toml interface of save.

Methods

impl Store

fn new(filename: &str) -> Result<Self>

The new constructor function returns a storage's interface.

fn get(&self) -> Option<(i64, i64)>

The get function returns a tuple of Theta zero and one.

fn set_zero(&mut self, zero: i64) -> Option<()>

The set_zero function updates the Theta zero.

fn set_one(&mut self, one: i64) -> Option<()>

The set_one function updates the Theta one.